Add module NuecaRailsInterface to interfaces
